use int64_t instead of long long

+++ contest/AOJ/Volume0/0019.cpp
@@ -1,7 +1,7 @@
 #include "template.hpp"
 
 int main() {
-  long long n, res = 1;
+  int64_t n, res = 1;
   cin >> n;
   for (int i = 1; i <= n; ++i) res *= i;
   cout << res << endl;
