+++ contest/AtCoder/abc029/D.cpp
@@ -1,8 +1,8 @@
 #include "template.hpp"
 
 int main() {
-  long long n, res = 0;
+  int64_t n, res = 0;
   cin >> n;
-  for (long long t = 1; t <= n; t *= 10) res += n / t / 10 * t + min(t, max(0ll, n % (t * 10) - t + 1));
+  for (int64_t t = 1; t <= n; t *= 10) res += n / t / 10 * t + min(t, max(int64_t(0), n % (t * 10) - t + 1));
   cout << res << endl;
 }
