+++ contest/AtCoder/code-festival-2014-relay/I.cpp
@@ -2,15 +2,15 @@
 #include "graph/edge.hpp"
 #include "graph/shortest_path.hpp"
 
-struct MyState : public DijkstraState<WeightedEdge<long long>> {
-  static vector<long long> a, b, c;
+struct MyState : public DijkstraState<WeightedEdge<int64_t>> {
+  static vector<int64_t> a, b, c;
 
-  MyState(const int pos) : DijkstraState<WeightedEdge<long long>>(pos) {}
+  MyState(const int pos) : DijkstraState<WeightedEdge<int64_t>>(pos) {}
 
-  MyState(const int from, const WeightedEdge<long long>& edge, long long cost) : DijkstraState<WeightedEdge<long long>>(from, edge, cost) {}
+  MyState(const int from, const WeightedEdge<int64_t>& edge, int64_t cost) : DijkstraState<WeightedEdge<int64_t>>(from, edge, cost) {}
 
-  MyState next(const int from, const WeightedEdge<long long>& edge) const {
-    long long k = cost;
+  MyState next(const int from, const WeightedEdge<int64_t>& edge) const {
+    int64_t k = cost;
     if (k < c[from]) {
       k = c[from];
     } else {
@@ -22,25 +22,25 @@ struct MyState : public DijkstraState<WeightedEdge<long long>> {
   }
 };
 
-vector<long long> MyState::a;
-vector<long long> MyState::b;
-vector<long long> MyState::c;
+vector<int64_t> MyState::a;
+vector<int64_t> MyState::b;
+vector<int64_t> MyState::c;
 
 int main() {
   int n, m, s, d;
   cin >> n >> m >> s >> d;
-  vector<long long> a(n), b(n), c(n);
+  vector<int64_t> a(n), b(n), c(n);
   for (int i = 0; i < n; ++i) cin >> a[i] >> b[i] >> c[i];
   MyState::a = a;
   MyState::b = b;
   MyState::c = c;
-  AdjacencyList<WeightedEdge<long long>> graph(n);
+  AdjacencyList<WeightedEdge<int64_t>> graph(n);
   for (int i = 0; i < m; ++i) {
     int x, y, t;
     cin >> x >> y >> t;
     graph.addUndirectedEdge(x - 1, y - 1, t);
   }
-  auto dijkstra = Dijkstra<AdjacencyList<WeightedEdge<long long>>, false, MyState>(graph);
+  auto dijkstra = Dijkstra<AdjacencyList<WeightedEdge<int64_t>>, false, MyState>(graph);
   dijkstra.solve(s - 1);
   cout << dijkstra.dis[d - 1] << endl;
 }
