+++ contest/AtCoder/indeednow-qualb/D.cpp
@@ -6,14 +6,14 @@ int main() {
   vector<int> a(n), p(c, -1);
   cin >> a;
   for (int& i : a) --i;
-  vector<long long> res(c, (long long)n * (n + 1) / 2);
+  vector<int64_t> res(c, (int64_t)n * (n + 1) / 2);
   for (int i = 0; i < n; ++i) {
-    long long d = i - p[a[i]];
+    int64_t d = i - p[a[i]];
     res[a[i]] -= d * (d - 1) / 2;
     p[a[i]] = i;
   }
   for (int i = 0; i < c; ++i) {
-    long long d = n - p[i];
+    int64_t d = n - p[i];
     res[i] -= d * (d - 1) / 2;
   }
   cout << res;
