+++ contest/AtCoder/snuke21/A.cpp
@@ -1,9 +1,9 @@
 #include "template.hpp"
 
 int main() {
-  long long n;
+  int64_t n;
   cin >> n;
-  for (long long i = sqrt(n); i * (i + 1) / 2 <= n; ++i) {
+  for (int64_t i = sqrt(n); i * (i + 1) / 2 <= n; ++i) {
     if (i * (i + 1) / 2 == n) {
       cout << i << endl;
       return 0;
