+++ contest/yukicoder/101.cpp
@@ -11,10 +11,10 @@ int main() {
     swap(a[x - 1], a[y - 1]);
   }
   vector<bool> used(n);
-  long long res = 1;
+  int64_t res = 1;
   for (int i = 0; i < n; ++i) {
     if (used[i]) continue;
-    long long c = 0;
+    int64_t c = 0;
     for (int j = i; !used[j]; j = a[j]) {
       used[j] = true;
       ++c;
