+++ contest/yukicoder/278.cpp
@@ -1,9 +1,9 @@
 #include "math/prime.hpp"
 
 int main() {
-  long long n;
+  int64_t n;
   cin >> n;
-  long long x = n % 2 ? n : n / 2, res = 0;
+  int64_t x = n % 2 ? n : n / 2, res = 0;
   Prime prime;
   for (auto& i : prime.divisor(x)) res += i;
   cout << res << endl;
