+++ contest/yukicoder/44.cpp
@@ -2,9 +2,9 @@
 #include "math/pow.hpp"
 
 int main() {
-  SquareMatrix<long long> m(2);
+  SquareMatrix<int64_t> m(2);
   m[0][0] = m[0][1] = m[1][0] = 1;
-  Vector<long long> v(2);
+  Vector<int64_t> v(2);
   v[0] = 1;
   int n;
   cin >> n;
