+++ contest/yukicoder/64.cpp
@@ -1,7 +1,7 @@
 #include "template.hpp"
 
 int main() {
-  long long f[3], n;
+  int64_t f[3], n;
   cin >> f[0] >> f[1] >> n;
   f[2] = f[0] ^ f[1];
   cout << f[n % 3] << endl;
