+++ contest/yukicoder/72.cpp
@@ -2,7 +2,7 @@
 
 int main() {
   Mint::setMod(1000007);
-  long long n;
+  int64_t n;
   cin >> n;
   Mint a = n / 2 + 1, b = (n + 1) / 2 + 1;
   cout << a * b - 1 << endl;
