+++ contest/yukicoder/84.cpp
@@ -1,7 +1,7 @@
 #include "template.hpp"
 
 int main() {
-  long long r, c;
+  int64_t r, c;
   cin >> r >> c;
   cout << r * c / (r == c ? 4 : 2) - (r * c % 2 ? 0 : 1) << endl;
 }
